diwali/requirements.txt
Django
google-generativeai
django-admin-sortable
